backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
